third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/historical/coep-navigate-popup-unsafe-inherit.https-expected.txt

This is a testharness.js-based test.
FAIL Popup navigating to same-origin with coop unsafe-inherit/coep assert_equals: opener expected true but got false
FAIL Popup navigating to same-site with coop unsafe-inherit/coep assert_equals: name expected "" but got "Popup-navigating-to-same-site-with-coop-unsafe-inherit/coep"
FAIL Popup navigating to same-origin with coop unsafe-inherit/no coep assert_equals: name expected "" but got "Popup-navigating-to-same-origin-with-coop-unsafe-inherit/no-coep"
FAIL Popup navigating to same-site with coop unsafe-inherit/no coep assert_equals: name expected "" but got "Popup-navigating-to-same-site-with-coop-unsafe-inherit/no-coep"
Harness: the test ran to completion.
